refactor(blockio): fold errors+requeues into blockio_requests sampler
Per @brayniac's review, avoid attaching a second BPF program at
block_rq_complete when blockio_requests already instruments it.

The new error and requeue counting paths now live inside the existing
blockio_requests sampler:

- block_rq_complete handler in blockio_requests/mod.bpf.c gets one
  added branch on non-OK status that classifies the blk_status_t into
  one of seven coarse buckets and bumps the per-CPU error slot. Hot
  path stays the same shape; the OK case falls through unchanged.

- A new block_rq_requeue program in the same BPF object handles the
  requeue counter. This is one more BPF program but at a tracepoint
  no other sampler instruments, so it's not duplicating work.

- All 32 new metric statics (28 errors + 4 requeues) move into
  blockio_requests/stats.rs alongside the existing blockio_operations
  / blockio_bytes / blockio_size declarations.

- The standalone blockio_errors sampler directory and its config /
  build.rs / mod.rs entries are removed.

Metric names (blockio_errors, blockio_requeues) are unchanged, so the
Exceptions dashboard queries continue to work without modification.

Build, clippy (clean of new warnings), fmt, and the workspace test
suite all pass.
