+++ config/agent.toml
@@ -62,17 +62,13 @@ enabled = true
 # Each sampler can then be individually configured to override the defaults. All
 # of the configuration options in the `[defaults]` section are allowed.
 
-# BPF sampler that instruments block_io completions and requeues to count
-# terminal IO failures (bucketed by op and error class) and recovered
-# requests put back on the queue.
-[samplers.blockio_errors]
-
 # BPF sampler that instruments block_io request queue to measure the request
 # latency distribution.
 [samplers.blockio_latency]
 
-# BPF sampler that instruments block_io request queue to get counts of requests,
-# the number of bytes by request type, and the size distribution.
+# BPF sampler that instruments block_io request queue. Produces per-op
+# completion counts, byte counters, size histograms, terminal-failure
+# counters bucketed by error class, and requeue counters.
 [samplers.blockio_requests]
 
 # BPF sampler that collects CPU CFS bandwidth control and throttling stats
