+++ src/agent/samplers/blockio/linux/mod.rs
@@ -1,3 +1,2 @@
-mod errors;
 mod latency;
 mod requests;
